Check isMounted in setTimeout
This fixes some occasional errors that I would see when causing a form
to unmount.  The unmount would happen before the setTimeout which then
caused setState to be called on an unmounted component.  React throws an
error in this case.

The fix is to check if the component is mounted before running the body
of the setTimeout